taxonomy-synthesis/pyproject.toml
[tool.poetry]
name = "taxonomy-synthesis"
version = "0.1.8"
description = "An AI-driven framework for synthesizing adaptive taxonomies, enabling automated data categorization and classification within dynamic hierarchical structures."
authors = ["Sebastian Sosa <1sebastian1sosa1@gmail.com> @ OPENGPT LLC <https://opengpt.llc/>"]
readme = "README.md"
packages = [{include = "taxonomy_synthesis"}]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^2.8.2"
openai = "^1.42.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pre-commit = "^3.8.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
